requirements.txt
appdirs==1.4.3
appnope==0.1.0
attrs==18.2.0
autopep8==1.4.3
backcall==0.1.0
black==18.9b0
Click==7.0
decorator==4.3.2
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.5.0
django-heroku==0.3.1
djangorestframework==3.9.2
gunicorn==19.9.0
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.3.1
python-decouple==3.1
pytz==2018.9
six==1.12.0
toml==0.10.0
traitlets==4.3.2
virtualenv==16.4.3
wcwidth==0.1.7
whitenoise==4.1.2